Memory leak was not a memory leak.
We were just blasting the ports harder than they could handle.  That is
fixed now, with a system that only makes another message once the
previous message has been successfully written.

All functionality in place before this refactoring is back.
Additionally, the dewd is more flexible in terms of allowable network
commands, and the io sessions are consolidated and more extendable.