Pre-Walter Commit.
The current plan is to have Walter exist as a mode of the dewd that is
available through configuration files.  Walter will be using some unix
specific headers, particularly for mkdir, just as the dewd used a unix
specific header for ioctl.  So the dewd isn't really portable outside of
unix systems.